move CheckCampaign provider
