+++ api/services/apdf/src/providers/CheckCampaign.ts
@@ -1,10 +1,10 @@
 import { KernelInterfaceResolver, provider } from '@ilos/common';
-import { handlerConfig } from '../../shared/apdf/export.contract';
+import { handlerConfig } from '../shared/apdf/export.contract';
 import {
   ParamsInterface as GetCampaignParamInterface,
   ResultInterface as GetCampaignResultInterface,
   signature as getCampaignSignature,
-} from '../../shared/policy/find.contract';
+} from '../shared/policy/find.contract';
 
 @provider()
 export class CheckCampaign {
